Implement new single input table view
