+++ Orange/widgets/data/owtable.py
@@ -4,9 +4,8 @@ import itertools
 import concurrent.futures
 
 from collections import namedtuple
-from typing import List, Optional
-
-from math import isnan
+from dataclasses import dataclass
+from typing import Optional, Union
 
 import numpy
 from scipy.sparse import issparse
@@ -18,10 +17,10 @@ from AnyQt.QtWidgets import (
 from AnyQt.QtGui import QColor, QClipboard
 from AnyQt.QtCore import (
     Qt, QSize, QEvent, QObject, QMetaObject,
-    QAbstractProxyModel, QIdentityProxyModel, QModelIndex,
-    QItemSelectionModel, QItemSelection, QItemSelectionRange,
+    QAbstractProxyModel, QItemSelectionModel, QItemSelection,
+    QItemSelectionRange,
 )
-from AnyQt.QtCore import pyqtSlot as Slot
+from AnyQt.QtCore import Slot
 
 import Orange.data
 from Orange.data.storage import Storage
@@ -30,6 +29,7 @@ from Orange.data.sql.table import SqlTable
 from Orange.statistics import basic_stats
 
 from Orange.widgets import gui
+from Orange.widgets.data.utils.models import RichTableModel, TableSliceProxy
 from Orange.widgets.settings import Setting
 from Orange.widgets.utils.itemdelegates import TableDataDelegate
 from Orange.widgets.utils.itemselectionmodel import (
@@ -38,7 +38,7 @@ from Orange.widgets.utils.itemselectionmodel import (
 from Orange.widgets.utils.tableview import TableView, \
     table_selection_to_mime_data
 from Orange.widgets.utils.widgetpreview import WidgetPreview
-from Orange.widgets.widget import OWWidget, MultiInput, Output, Msg
+from Orange.widgets.widget import OWWidget, Input, Output, Msg
 from Orange.widgets.utils import datacaching
 from Orange.widgets.utils.localization import pl
 from Orange.widgets.utils.annotated_data import (create_annotated_table,
@@ -47,134 +47,19 @@ from Orange.widgets.utils.itemmodels import TableModel
 from Orange.widgets.utils.state_summary import format_summary_details
 
 
-class RichTableModel(TableModel):
-    """A TableModel with some extra bells and whistles/
-
-    (adds support for gui.BarRole, include variable labels and icons
-    in the header)
-    """
-    #: Rich header data flags.
-    Name, Labels, Icon = 1, 2, 4
-
-    def __init__(self, sourcedata, parent=None):
-        super().__init__(sourcedata, parent)
-
-        self._header_flags = RichTableModel.Name
-        self._continuous = [var.is_continuous for var in self.vars]
-        labels = []
-        for var in self.vars:
-            if isinstance(var, Orange.data.Variable):
-                labels.extend(var.attributes.keys())
-        self._labels = list(sorted(
-            {label for label in labels if not label.startswith("_")}))
-
-    def data(self, index, role=Qt.DisplayRole,
-             # for faster local lookup
-             _BarRole=gui.TableBarItem.BarRole):
-        # pylint: disable=arguments-differ
-        if role == _BarRole and self._continuous[index.column()]:
-            val = super().data(index, TableModel.ValueRole)
-            if val is None or isnan(val):
-                return None
-
-            dist = super().data(index, TableModel.VariableStatsRole)
-            if dist is not None and dist.max > dist.min:
-                return (val - dist.min) / (dist.max - dist.min)
-            else:
-                return None
-        elif role == Qt.TextAlignmentRole and self._continuous[index.column()]:
-            return Qt.AlignRight | Qt.AlignVCenter
-        else:
-            return super().data(index, role)
-
-    def headerData(self, section, orientation, role):
-        if orientation == Qt.Horizontal and role == Qt.DisplayRole:
-            var = super().headerData(
-                section, orientation, TableModel.VariableRole)
-            if var is None:
-                return super().headerData(
-                    section, orientation, Qt.DisplayRole)
-
-            lines = []
-            if self._header_flags & RichTableModel.Name:
-                lines.append(var.name)
-            if self._header_flags & RichTableModel.Labels:
-                lines.extend(str(var.attributes.get(label, ""))
-                             for label in self._labels)
-            return "\n".join(lines)
-        elif orientation == Qt.Horizontal and role == Qt.DecorationRole and \
-                self._header_flags & RichTableModel.Icon:
-            var = super().headerData(
-                section, orientation, TableModel.VariableRole)
-            if var is not None:
-                return gui.attributeIconDict[var]
-            else:
-                return None
-        else:
-            return super().headerData(section, orientation, role)
-
-    def setRichHeaderFlags(self, flags):
-        if flags != self._header_flags:
-            self._header_flags = flags
-            self.headerDataChanged.emit(
-                Qt.Horizontal, 0, self.columnCount() - 1)
-
-    def richHeaderFlags(self):
-        return self._header_flags
-
-
-class TableSliceProxy(QIdentityProxyModel):
-    def __init__(self, parent=None, rowSlice=slice(0, -1), **kwargs):
-        super().__init__(parent, **kwargs)
-        self.__rowslice = rowSlice
-
-    def setRowSlice(self, rowslice):
-        if rowslice.step is not None and rowslice.step != 1:
-            raise ValueError("invalid stride")
-
-        if self.__rowslice != rowslice:
-            self.beginResetModel()
-            self.__rowslice = rowslice
-            self.endResetModel()
-
-    def mapToSource(self, proxyindex):
-        model = self.sourceModel()
-        if model is None or not proxyindex.isValid():
-            return QModelIndex()
-
-        row, col = proxyindex.row(), proxyindex.column()
-        row = row + self.__rowslice.start
-        assert 0 <= row < model.rowCount()
-        return model.createIndex(row, col, proxyindex.internalPointer())
-
-    def mapFromSource(self, sourceindex):
-        model = self.sourceModel()
-        if model is None or not sourceindex.isValid():
-            return QModelIndex()
-        row, col = sourceindex.row(), sourceindex.column()
-        row = row - self.__rowslice.start
-        assert 0 <= row < self.rowCount()
-        return self.createIndex(row, col, sourceindex.internalPointer())
-
-    def rowCount(self, parent=QModelIndex()):
-        if parent.isValid():
-            return 0
-        count = super().rowCount()
-        start, stop, step = self.__rowslice.indices(count)
-        assert step == 1
-        return stop - start
-
-
-TableSlot = namedtuple("TableSlot", ["input_id", "table", "summary", "view"])
-
-
 class DataTableView(gui.HScrollStepMixin, TableView):
     dataset: Table
-    input_slot: TableSlot
 
 
 class TableBarItemDelegate(gui.TableBarItem, TableDataDelegate):
     pass
+
+
+@dataclass
+class InputData:
+    table: Table
+    summary: Union['Summary', 'ApproxSummary']
+    model: TableModel
 
 
 class OWDataTable(OWWidget):
@@ -185,22 +70,15 @@ class OWDataTable(OWWidget):
     keywords = "data table, view"
 
     class Inputs:
-        data = MultiInput("Data", Table, auto_summary=False, filter_none=True)
+        data = Input("Data", Table)
 
     class Outputs:
         selected_data = Output("Selected Data", Table, default=True)
         annotated_data = Output(ANNOTATED_DATA_SIGNAL_NAME, Table)
 
-    class Warning(OWWidget.Warning):
-        multiple_inputs = Msg(
-            "Multiple Data inputs are deprecated.\n"
-            "This functionality will be removed soon.\n"
-            "Use multiple Tables instead.")
-
     buttons_area_orientation = Qt.Vertical
 
     show_distributions = Setting(False)
-    dist_color_RGB = Setting((220, 220, 220, 255))
     show_attribute_labels = Setting(True)
     select_rows = Setting(True)
     auto_commit = Setting(True)
@@ -209,27 +87,25 @@ class OWDataTable(OWWidget):
     selected_rows = Setting([], schema_only=True)
     selected_cols = Setting([], schema_only=True)
 
-    settings_version = 2
+    settings_version = 1
 
     def __init__(self):
         super().__init__()
-        self._inputs: List[TableSlot] = []
+        self.input: Optional[InputData] = None
         self.__pending_selected_rows = self.selected_rows
         self.selected_rows = None
         self.__pending_selected_cols = self.selected_cols
         self.selected_cols = None
-
-        self.dist_color = QColor(*self.dist_color_RGB)
+        self.dist_color = QColor(220, 220, 220, 255)
 
         info_box = gui.vBox(self.controlArea, "Info")
         self.info_text = gui.widgetLabel(info_box)
-        self._set_input_summary(None)
 
         box = gui.vBox(self.controlArea, "Variables")
         self.c_show_attribute_labels = gui.checkBox(
             box, self, "show_attribute_labels",
             "Show variable labels (if present)",
-            callback=self._on_show_variable_labels_changed)
+            callback=self._update_variable_labels)
 
         gui.checkBox(box, self, "show_distributions",
                      'Visualize numeric values',
@@ -251,19 +127,9 @@ class OWDataTable(OWWidget):
                    attribute=Qt.WA_LayoutUsesWidgetRect)
         gui.auto_send(self.buttonsArea, self, "auto_commit")
 
-        # GUI with tabs
-        self.tabs = gui.tabWidget(self.mainArea)
-        self.tabs.currentChanged.connect(self._on_current_tab_changed)
-
-    def copy_to_clipboard(self):
-        self.copy()
-
-    @staticmethod
-    def sizeHint():
-        return QSize(800, 500)
-
-    def _create_table_view(self):
-        view = DataTableView()
+        view = DataTableView(
+            sortingEnabled=True
+        )
         view.setSortingEnabled(True)
         view.setItemDelegate(TableDataDelegate(view))
 
@@ -275,6 +141,7 @@ class OWDataTable(OWWidget):
         header.setSectionsClickable(True)
         header.setSortIndicatorShown(True)
         header.setSortIndicator(-1, Qt.AscendingOrder)
+        header.sortIndicatorChanged.connect(self.update_selection)
 
         # QHeaderView does not 'reset' the model sort column,
         # because there is no guaranty (requirement) that the
@@ -282,74 +149,45 @@ class OWDataTable(OWWidget):
         def sort_reset(index, order):
             if view.model() is not None and index == -1:
                 view.model().sort(index, order)
+
         header.sortIndicatorChanged.connect(sort_reset)
-        return view
+        self.view = view
+        self.mainArea.layout().addWidget(self.view)
+        self._update_input_summary()
+
+    def copy_to_clipboard(self):
+        self.copy()
+
+    def sizeHint(self):
+        return QSize(800, 500)
 
     @Inputs.data
-    def set_dataset(self, index: int, data: Table):
+    def set_dataset(self, data: Optional[Table]):
         """Set the input dataset."""
-        datasetname = getattr(data, "name", "Data")
-        slot = self._inputs[index]
-        view = slot.view
         # reset the (header) view state.
-        view.setModel(None)
-        view.horizontalHeader().setSortIndicator(-1, Qt.AscendingOrder)
-        assert self.tabs.indexOf(view) != -1
-        self.tabs.setTabText(self.tabs.indexOf(view), datasetname)
-        view.dataset = data
-        slot = TableSlot(index, data, table_summary(data), view)
-        view.input_slot = slot
-        self._inputs[index] = slot
-        self._setup_table_view(view, data)
-        self.tabs.setCurrentWidget(view)
-        self._set_multi_input_warning()
-
-    @Inputs.data.insert
-    def insert_dataset(self, index: int, data: Table):
-        datasetname = getattr(data, "name", "Data")
-        view = self._create_table_view()
-        slot = TableSlot(None, data, table_summary(data), view)
-        view.dataset = data
-        view.input_slot = slot
-        self._inputs.insert(index, slot)
-        self.tabs.insertTab(index, view, datasetname)
-        self._setup_table_view(view, data)
-        self.tabs.setCurrentWidget(view)
-        self._set_multi_input_warning()
-
-    @Inputs.data.remove
-    def remove_dataset(self, index):
-        slot = self._inputs.pop(index)
-        view = slot.view
-        self.tabs.removeTab(self.tabs.indexOf(view))
-        view.setModel(None)
-        view.hide()
-        view.deleteLater()
-
-        current = self.tabs.currentWidget()
-        if current is not None:
-            self._set_input_summary(current.input_slot)
-        self._set_multi_input_warning()
-
-    def _set_multi_input_warning(self):
-        self.Warning.multiple_inputs(shown=len(self._inputs) > 1)
+        self.view.setModel(None)
+        self.view.horizontalHeader().setSortIndicator(-1, Qt.AscendingOrder)
+        if data is not None:
+            self.input = InputData(
+                table=data,
+                summary=table_summary(data),
+                model=RichTableModel(data)
+            )
+            self._setup_table_view()
+        else:
+            self.input = None
 
     def handleNewSignals(self):
         super().handleNewSignals()
-        self.tabs.tabBar().setVisible(self.tabs.count() > 1)
-        data: Optional[Table] = None
-        current = self.tabs.currentWidget()
-        slot = None
-        if current is not None:
-            data = current.dataset
-            slot = current.input_slot
-
-        if slot and isinstance(slot.summary.len, concurrent.futures.Future):
+        data: Optional[Table] = self.input.table if self.input else None
+        slot = self.input
+        if slot is not None and isinstance(slot.summary.len, concurrent.futures.Future):
             def update(_):
                 QMetaObject.invokeMethod(
                     self, "_update_info", Qt.QueuedConnection)
             slot.summary.len.add_done_callback(update)
-        self._set_input_summary(slot)
+
+        self._update_input_summary()
 
         if data is not None and self.__pending_selected_rows is not None:
             self.selected_rows = self.__pending_selected_rows
@@ -366,10 +204,15 @@ class OWDataTable(OWWidget):
         self.set_selection()
         self.commit.now()
 
-    def _setup_table_view(self, view, data):
-        """Setup the `view` (QTableView) with `data` (Orange.data.Table)
-        """
-        datamodel = RichTableModel(data)
+    def _setup_table_view(self):
+        """Setup the view with current input data."""
+        if self.input is None:
+            self.view.setModel(None)
+            return
+
+        datamodel = self.input.model
+        view = self.view
+        data = self.input.table
         rowcount = data.approx_len()
 
         if self.color_by_class and data.domain.has_discrete_class:
@@ -390,7 +233,6 @@ class OWDataTable(OWWidget):
         header = view.horizontalHeader()
         header.setSectionsClickable(is_sortable(data))
         header.setSortIndicatorShown(is_sortable(data))
-        header.sortIndicatorChanged.connect(self.update_selection)
 
         view.setModel(datamodel)
 
@@ -420,7 +262,7 @@ class OWDataTable(OWWidget):
         assert vheader.sectionSize(0) > 1 or datamodel.rowCount() == 0
 
         # update the header (attribute names)
-        self._update_variable_labels(view)
+        self._update_variable_labels()
 
         selmodel = BlockSelectionModel(
             view.model(), parent=view, selectBlocks=not self.select_rows)
@@ -438,8 +280,7 @@ class OWDataTable(OWWidget):
                 btn = table.findChild(QAbstractButton)
 
                 class Efc(QObject):
-                    @staticmethod
-                    def eventFilter(o, e):
+                    def eventFilter(self, o, e):
                         if (isinstance(o, QAbstractButton) and
                                 e.type() == QEvent.Paint):
                             # paint by hand (borrowed from QTableCornerButton)
@@ -490,7 +331,7 @@ class OWDataTable(OWWidget):
             except Exception:
                 pass
 
-    def _set_input_summary(self, slot):
+    def _update_input_summary(self):
         def format_summary(summary):
             if isinstance(summary, ApproxSummary):
                 length = summary.len.result() if summary.len.done() else \
@@ -500,12 +341,12 @@ class OWDataTable(OWWidget):
             return length
 
         summary, details = self.info.NoInput, ""
-        if slot:
-            summary = format_summary(slot.summary)
-            details = format_summary_details(slot.table)
+        if self.input:
+            summary = format_summary(self.input.summary)
+            details = format_summary_details(self.input.table)
         self.info.set_input_summary(summary, details)
 
-        self.info_text.setText("\n".join(self._info_box_text(slot)))
+        self.info_text.setText("\n".join(self._info_box_text(self.input)))
 
     @staticmethod
     def _info_box_text(slot):
@@ -568,32 +409,23 @@ class OWDataTable(OWWidget):
         return text
 
     def _on_select_all(self, _):
-        data_info = self.tabs.currentWidget().input_slot.summary
+        data_info = self.input.summary
         if len(self.selected_rows) == data_info.len \
                 and len(self.selected_cols) == len(data_info.domain.variables):
-            self.tabs.currentWidget().clearSelection()
+            self.view.clearSelection()
         else:
-            self.tabs.currentWidget().selectAll()
+            self.view.selectAll()
 
-    def _on_current_tab_changed(self, index):
-        """Update the status bar on current tab change"""
-        view = self.tabs.widget(index)
-        if view is not None and view.model() is not None:
-            self._set_input_summary(view.input_slot)
-            self.update_selection()
-        else:
-            self._set_input_summary(None)
-
-    def _update_variable_labels(self, view):
-        "Update the variable labels visibility for `view`"
-        model = view.model()
-        if isinstance(model, TableSliceProxy):
-            model = model.sourceModel()
-
+    def _update_variable_labels(self):
+        """Update the variable labels visibility for current view."""
+        view = self.view
+        if self.input is None:
+            return
+        model = self.input.model
         if self.show_attribute_labels:
             model.setRichHeaderFlags(
-                RichTableModel.Labels | RichTableModel.Name)
-
+                RichTableModel.Labels | RichTableModel.Name
+            )
             labelnames = set()
             domain = model.source.domain
             for a in itertools.chain(domain.metas, domain.variables):
@@ -605,65 +437,54 @@ class OWDataTable(OWWidget):
             model.setRichHeaderFlags(RichTableModel.Name)
             self.set_corner_text(view, "")
 
-    def _on_show_variable_labels_changed(self):
-        """The variable labels (var.attribues) visibility was changed."""
-        for slot in self._inputs:
-            self._update_variable_labels(slot.view)
-
     def _on_distribution_color_changed(self):
-        for ti in range(self.tabs.count()):
-            widget = self.tabs.widget(ti)
-            model = widget.model()
-            while isinstance(model, QAbstractProxyModel):
-                model = model.sourceModel()
-            data = model.source
-            class_var = data.domain.class_var
-            if self.color_by_class and class_var and class_var.is_discrete:
-                color_schema = [QColor(*c) for c in class_var.colors]
-            else:
-                color_schema = None
-            if self.show_distributions:
-                delegate = TableBarItemDelegate(widget, color=self.dist_color,
-                                                color_schema=color_schema)
-            else:
-                delegate = TableDataDelegate(widget)
-            widget.setItemDelegate(delegate)
-        tab = self.tabs.currentWidget()
-        if tab:
-            tab.reset()
+        if self.input is None:
+            return
+        widget = self.view
+        model = self.input.model
+        data = model.source
+        class_var = data.domain.class_var
+        if self.color_by_class and class_var and class_var.is_discrete:
+            color_schema = [QColor(*c) for c in class_var.colors]
+        else:
+            color_schema = None
+        if self.show_distributions:
+            delegate = TableBarItemDelegate(widget, color=self.dist_color,
+                                            color_schema=color_schema)
+        else:
+            delegate = TableDataDelegate(widget)
+        widget.setItemDelegate(delegate)
+        self.view.reset()
 
     def _on_select_rows_changed(self):
-        for slot in self._inputs:
-            selection_model = slot.view.selectionModel()
-            selection_model.setSelectBlocks(not self.select_rows)
-            if self.select_rows:
-                slot.view.setSelectionBehavior(QTableView.SelectRows)
-                # Expand the current selection to full row selection.
-                selection_model.select(
-                    selection_model.selection(),
-                    QItemSelectionModel.Select | QItemSelectionModel.Rows
-                )
-            else:
-                slot.view.setSelectionBehavior(QTableView.SelectItems)
+        if self.input is None:
+            return
+        selection_model = self.view.selectionModel()
+        selection_model.setSelectBlocks(not self.select_rows)
+        if self.select_rows:
+            self.view.setSelectionBehavior(QTableView.SelectRows)
+            # Expand the current selection to full row selection.
+            selection_model.select(
+                selection_model.selection(),
+                QItemSelectionModel.Select | QItemSelectionModel.Rows
+            )
+        else:
+            self.view.setSelectionBehavior(QTableView.SelectItems)
 
     def restore_order(self):
         """Restore the original data order of the current view."""
-        table = self.tabs.currentWidget()
-        if table is not None:
-            table.horizontalHeader().setSortIndicator(-1, Qt.AscendingOrder)
+        self.view.horizontalHeader().setSortIndicator(-1, Qt.AscendingOrder)
 
     @Slot()
     def _update_info(self):
-        current = self.tabs.currentWidget()
-        if current is not None and current.model() is not None:
-            self._set_input_summary(current.input_slot)
+        self._update_input_summary()
 
     def update_selection(self, *_):
         self.commit.deferred()
 
     def set_selection(self):
         if self.selected_rows and self.selected_cols:
-            view = self.tabs.currentWidget()
+            view = self.view
             model = view.model()
             if model.rowCount() <= self.selected_rows[-1] or \
                     model.columnCount() <= self.selected_cols[-1]:
@@ -710,23 +531,16 @@ class OWDataTable(OWWidget):
         rows = rows.tolist()
         return rows, cols
 
-    @staticmethod
-    def _get_model(view):
-        model = view.model()
-        while isinstance(model, QAbstractProxyModel):
-            model = model.sourceModel()
-        return model
-
     @gui.deferred
     def commit(self):
         """
         Commit/send the current selected row/column selection.
         """
         selected_data = table = rowsel = None
-        view = self.tabs.currentWidget()
-        if view and view.model() is not None:
-            model = self._get_model(view)
-            table = model.source  # The input data table
+        view = self.view
+        if self.input is not None:
+            model = self.input.model
+            table = self.input.table
 
             # Selections of individual instances are not implemented
             # for SqlTables
@@ -773,20 +587,18 @@ class OWDataTable(OWWidget):
         """
         Copy current table selection to the clipboard.
         """
-        view = self.tabs.currentWidget()
-        if view is not None:
-            mime = table_selection_to_mime_data(view)
+        if self.input is not None:
+            mime = table_selection_to_mime_data(self.view)
             QApplication.clipboard().setMimeData(
                 mime, QClipboard.Clipboard
             )
 
     def send_report(self):
-        view = self.tabs.currentWidget()
-        if not view or not view.model():
+        if self.input is None:
             return
-        model = self._get_model(view)
+        model = self.input.model
         self.report_data_brief(model.source)
-        self.report_table(view)
+        self.report_table(self.view)
 
 
 # Table Summary
@@ -849,9 +661,7 @@ def table_summary(table):
             elif density == Storage.MISSING:
                 return NotAvailable()
             else:
-                assert False
-                return None
-
+                raise ValueError
         X_part = parts(table.X, table.X_density(), X_dist)
         Y_part = parts(table.Y, table.Y_density(), Y_dist)
         M_part = parts(table.metas, table.metas_density(), M_dist)
@@ -869,8 +679,5 @@ def is_sortable(table):
 
 if __name__ == "__main__":  # pragma: no cover
     WidgetPreview(OWDataTable).run(
-        insert_dataset=[
-            (0, Table("iris")),
-            (1, Table("brown-selected")),
-            (2, Table("housing"))
-        ])
+        input_data=Table("iris"),
+    )
